Move type Model and Module into module.go
