+++ optim.go
@@ -11,6 +11,11 @@ import (
 	"unsafe"
 )
 
+// Optimizer struct
+type Optimizer struct {
+	Opt *C.Optimizer
+}
+
 // NewSGDOpt creates a SGD Optimizer
 func NewSGDOpt(lr, momentum, dampening, weightDecay float64, nesterov bool) Optimizer {
 	nt := 0
